Extracted the `ServiceConfigurationImpl` class from the file `ServiceConfiguration`.
